pass onetoone get klass

+++ src/main/java/com/oocl/jpapractice/onetoone/repositories/KlassRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KlassRepository extends JpaRepository<Klass, Long> {
+public interface KlassRepository extends JpaRepository<Klass, Integer> {
 }
