+++ src/main/java/com/oocl/jpapractice/onetoone/repositories/LeaderRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LeaderRepository extends JpaRepository<Leader, Long> {
+public interface LeaderRepository extends JpaRepository<Leader, Integer> {
 }
